GhostTargetWindow to seperate file
